Move test related helpers to seperate folder
