+++ tests/task.resolvers.mutation.e2e.test.js
@@ -1,4 +1,4 @@
-const apolloHttpServer = require("../src/helpers/mockApolloHttpServer");
+const apolloHttpServer = require("../tests_helpers/mockApolloHttpServer");
 const request = require("supertest");
 const context = require("../src/context");
 const { jwt } = require("../src/helpers");
